fix(ci): splitted diff methods into seperate files and fixed linting issues
